+++ view/src/utils/error_printer.rs
@@ -1,3 +1,6 @@
-pub fn error_printer(error: std::io::Error){
+use std::error::Error;
+
+pub fn error_printer(error: Box<dyn Error>){
+
     eprintln!("Error : {:?}",error);
 }
