+++ view/src/utils/request_receiver.rs
@@ -1,6 +1,7 @@
+use std::error::Error;
 use std::io;
 
-pub fn request_receiver() -> Result<String,io::Error> {
+pub fn request_receiver() -> Result<String,Box<dyn Error>> {
     println!("Enter a SQL request :");
 
     let mut request = String::new();
